backend/pyproject.toml
[tool.poetry]
name = "time-glass-backend"
version = "0.1.0"
description = "Time Glass Backend API"
authors = ["Your Name"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.27.0"
sqlalchemy = "^2.0.25"
pydantic = "^2.5.3"
python-dotenv = "^1.0.0"
alembic = "^1.13.1"
elasticsearch = "^8.11.1"
pydantic-settings = "^2.1.0"
uuid = "^1.30"
aiohttp = "^3.9.3"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
